refactor: 게시글 쓰기, 수정 컴포넌트 생성
